Removed outdated file, fixed improper filename in get_key() method
